Separate functionality into walinator
